Refactor client into modular scripts
