+++ Client/stream.py
@@ -1,171 +1,28 @@
+# -*- coding: utf-8 -*-
+"""Camera streaming and barcode detection loop."""
+
+from __future__ import annotations
 
 import os
-import argparse
-import urllib.request
-import time
 from typing import Dict
 
 import cv2
-import requests
 from pyzbar import pyzbar
+
+from utils import send_barcode
+
 try:
     from ultralytics import YOLO
 except ImportError:  # pragma: no cover - optional dependency
     YOLO = None  # type: ignore
 
 
-def ensure_wechat_models(det_proto: str, det_model: str, sr_proto: str, sr_model: str) -> None:
-    """Download WeChatQRCode models if they don't exist."""
-    base_url = "https://raw.githubusercontent.com/WeChatCV/opencv_3rdparty/wechat_qrcode/"
-    for path in [det_proto, det_model, sr_proto, sr_model]:
-        if not os.path.exists(path):
-            url = base_url + os.path.basename(path)
-            try:
-                print(f"Downloading {url} ...")
-                urllib.request.urlretrieve(url, path)
-            except Exception as exc:  # pragma: no cover - network issues
-                print(f"Failed to download {url}: {exc}")
-
-
-def parse_camera_info(info_path: str) -> Dict[str, str]:
-    """Parse camera metadata from the info file."""
-    info: Dict[str, str] = {}
-    if os.path.exists(info_path):
-        with open(info_path, "r", encoding="utf-8") as f:
-            for line in f:
-                if ":" in line:
-                    key, value = line.strip().split(":", 1)
-                    info[key.strip()] = value.strip()
-    return info
-
-
-def send_barcode(data: str, info: Dict[str, str], last: Dict[str, float]) -> None:
-    """Send scanned barcode data to the server.
-
-    A barcode will only be posted if at least two seconds have passed since the
-    last time that same barcode was sent. This prevents repeatedly hitting the
-    server when the same code is scanned in rapid succession.
-    """
-    if not data:
-        return
-
-    now = time.time()
-
-    # Only send if more than two seconds have passed since this barcode was
-    # last posted.
-    if now - last.get(data, 0) < 2:
-        return
-
-    payload = {
-        "barcode": data,
-        "camera_name": info.get("Camera Name", ""),
-        "area": info.get("Camera Area", ""),
-        "camera_type": info.get("Camera Type", ""),
-        "client_ip": info.get("Client IP", ""),
-        "camera_url": info.get("Camera URL", ""),
-    }
-    url = f"http://{info.get('Server IP', 'localhost')}:{info.get('Port', '5000')}/api/scan"
-
-    try:
-        requests.post(url, json=payload, timeout=2)
-    except Exception as exc:  # pragma: no cover - network or server errors
-        print(f"Failed to send barcode data: {exc}")
-
-    # Record the time this barcode was sent
-    last[data] = now
-
-
-def main() -> None:
-    """Start the webcam stream and scan for barcodes or QR codes."""
-    parser = argparse.ArgumentParser(
-        description="Start the webcam stream and scan for barcodes or QR codes."
-    )
-    parser.add_argument(
-        "--model-path",
-        default=os.getenv("YOLO_MODEL_PATH", "barcode_yolo.pt"),
-        help="Path to the YOLO model for barcode detection",
-    )
-    parser.add_argument(
-        "--wechat-det-prototxt",
-        default=os.getenv("WECHAT_DET_PROTOTXT", "detect.prototxt"),
-        help="Path to WeChat QRCode detection prototxt file",
-    )
-    parser.add_argument(
-        "--wechat-det-model",
-        default=os.getenv("WECHAT_DET_MODEL", "detect.caffemodel"),
-        help="Path to WeChat QRCode detection model file",
-    )
-    parser.add_argument(
-        "--wechat-sr-prototxt",
-        default=os.getenv("WECHAT_SR_PROTOTXT", "sr.prototxt"),
-        help="Path to WeChat QRCode super resolution prototxt file",
-    )
-    parser.add_argument(
-        "--wechat-sr-model",
-        default=os.getenv("WECHAT_SR_MODEL", "sr.caffemodel"),
-        help="Path to WeChat QRCode super resolution model file",
-    )
-
-    args = parser.parse_args()
-
-    # Load previous camera metadata if available and prompt for updates
-    info_path = os.path.join(os.path.dirname(__file__), "camera_info.txt")
-    prev_info = parse_camera_info(info_path)
-
-    camera_name = input(
-        f"Camera Name [{prev_info.get('Camera Name', '')}]: "
-    ) or prev_info.get("Camera Name", "")
-    camera_area = input(
-        f"Camera Area [{prev_info.get('Camera Area', '')}]: "
-    ) or prev_info.get("Camera Area", "")
-    camera_type = input(
-        f"Camera Type [{prev_info.get('Camera Type', '')}]: "
-    ) or prev_info.get("Camera Type", "")
-    server_IP = input(
-        f"Server IP [{prev_info.get('Server IP', '')}]: "
-    ) or prev_info.get("Server IP", "")
-    port = input(f"Port [{prev_info.get('Port', '')}]: ") or prev_info.get("Port", "")
-    client_ip = input(
-        f"Client IP [{prev_info.get('Client IP', '')}]: "
-    ) or prev_info.get("Client IP", "")
-    camera_url_input = input(
-        f"Camera URL [{prev_info.get('Camera URL', '')}]: "
-    ) or prev_info.get("Camera URL", "")
-
-    with open(info_path, "w", encoding="utf-8") as f:
-        f.write(f"Camera Name: {camera_name}\n")
-        f.write(f"Camera Area: {camera_area}\n")
-        f.write(f"Camera Type: {camera_type}\n")
-        f.write(f"Server IP: {server_IP}\n")
-        f.write(f"Port: {port}\n")
-        f.write(f"Client IP: {client_ip}\n")
-        f.write(f"Camera URL: {camera_url_input}\n")
-
-    print(f"Camera info saved to {info_path}")
-
-    camera_info = parse_camera_info(info_path)
-    # Tracks the last time each barcode was posted
+def run_stream(cap: cv2.VideoCapture, camera_info: Dict[str, str], args) -> None:
+    """Display ``cap`` frames and detect barcodes/QR codes."""
     last_scanned: Dict[str, float] = {}
 
-    ensure_wechat_models(
-        args.wechat_det_prototxt,
-        args.wechat_det_model,
-        args.wechat_sr_prototxt,
-        args.wechat_sr_model,
-    )
-    camera_url = camera_info.get("Camera URL", "0")
-    cam_source = int(camera_url) if camera_url.isdigit() else camera_url
-    cap = cv2.VideoCapture(cam_source)
-    if not cap.isOpened():
-        print(f"Cannot open camera {camera_url}")
-        return
-
-    # Try to capture at a higher resolution to help with far away codes
-    cap.set(cv2.CAP_PROP_FRAME_WIDTH, 1280)
-    cap.set(cv2.CAP_PROP_FRAME_HEIGHT, 1080)
-
     detector = cv2.QRCodeDetector()
-    barcode_detector = cv2.barcode_BarcodeDetector() # type: ignore
+    barcode_detector = cv2.barcode_BarcodeDetector()  # type: ignore
 
     wechat_detector = None
     if hasattr(cv2, "wechat_qrcode"):
@@ -179,14 +36,13 @@ def main() -> None:
         except Exception as exc:  # pragma: no cover - optional dependency
             print(f"Failed to load WeChatQRCode models: {exc}")
 
-    # Load YOLO model for barcode detection if available
     yolo_model = None
     if YOLO is not None:
         model_path = args.model_path
         if os.path.exists(model_path):
             try:
                 yolo_model = YOLO(model_path)
-            except Exception as exc:  # pragma: no cover - model loading is optional
+            except Exception as exc:  # pragma: no cover - model loading optional
                 print(f"Failed to load YOLO model from {model_path}: {exc}")
         else:
             print(
@@ -263,7 +119,7 @@ def main() -> None:
                     barcodes = pyzbar.decode(gray)
                     if not barcodes and yolo_model is not None:
                         results = yolo_model(frame, verbose=False)
-                        for bbox in results.boxes.xyxy.tolist(): # type: ignore
+                        for bbox in results.boxes.xyxy.tolist():  # type: ignore
                             x1, y1, x2, y2 = map(int, bbox)
                             roi = gray[y1:y2, x1:x2]
                             detected = pyzbar.decode(roi)
@@ -312,7 +168,3 @@ def main() -> None:
 
     cap.release()
     cv2.destroyAllWindows()
-
-
-if __name__ == '__main__':
-    main()
