+++ README.md
@@ -1,2 +1,20 @@
 # WareEye
 
+## Usage
+
+### Server
+
+```bash
+cd Server
+python app.py
+```
+
+### Client
+
+```bash
+cd Client
+python client.py
+```
+
+The client prompts for camera details and streams video while scanning for barcodes.
+
